+++ src/core/shared/utility/undefinedCheck.ts
@@ -1,0 +1,11 @@
+export function isNullOrUndefined(value: any) {
+    if (typeof value === 'undefined') {
+        return true;
+    }
+
+    if (typeof value !== 'number' && typeof value !== 'boolean' && !value) {
+        return true;
+    }
+
+    return false;
+}